requirements.txt
streamlit
pandas
numpy
scikit-learn
matplotlib
seaborn
sklearn-genetic-opt
# Optional:
# xgboost
# lightgbm
